fixed data posting info, also split functions to different files
