+++ src/pages/NewPost.js
@@ -1,6 +1,4 @@
 import React from 'react';
-import { Link } from 'react-router-dom';
-import ReactDOM from "react-dom";
 
 class NewPost extends React.Component {
